+++ src/FNO.Domain/Events/Market/OrderPartiallyFulfilledEvent.cs
@@ -4,8 +4,8 @@ namespace FNO.Domain.Events.Market
 {
     public class OrderPartiallyFulfilledEvent : EntityEvent
     {
-        public int QuantityFulfilled { get; set; }
-        public int Price { get; set; }
+        public long QuantityFulfilled { get; set; }
+        public long Price { get; set; }
 
         public OrderPartiallyFulfilledEvent()
         {
